[#6] Refactor: container 모듈에 init render 위임, 파일구조 변경
- main 파일에서 이벤트 리스너 등록하도록 변경
