+++ FE/src/ts/components/fetch.ts
@@ -1,6 +1,6 @@
 import { Card } from './card';
 import { createCardElement } from './card';
-import { initialRender } from './column';
+import { initialRender } from './container';
 
 const BASE_URL: string = 'https://576272fa-2ef9-48d0-a2c7-8ff6e25f9352.mock.pstmn.io';
