+++ FE/src/ts/main.ts
@@ -1,3 +1,7 @@
 import { fetchTodoList } from './components/fetch';
+import { clickHandler, inputHandler } from './components/eventManager';
+import { columnWrapElement } from './components/container';
 
 fetchTodoList();
+columnWrapElement.addEventListener('click', clickHandler);
+columnWrapElement.addEventListener('input', inputHandler);
