+++ FE/src/ts/utils/constants.ts
@@ -8,5 +8,3 @@ export const UTIL_CLASS = {
   hidden: 'hidden',
   disabled: 'disabled',
 };
-
-export const WRAPPER_CLASS: string = 'column-wrap';
